CMTOOL-168: updates elytron subsystem configurations added to server configurations migrated to EAP 7.1
